Updated values and corrected mathematics
I updated the value of diceThrows to be equal to the minimum number I
was able to repeatedly get experimental results within 4 sigma of the
exact calculated results.

I also updated the math so that the error terms were accurate.
Previously the error would be calculated by taking the absolute value
of the result of a subtraction, where in reality what returns the
correct amount is taking the absolute value of each terms before
subtracting them and then taking the absolute value again.